% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zero.R
\name{add_zero}
\alias{add_zero}
\title{Prepend zero(s) to a number}
\usage{
add_zero(x, n)
}
\arguments{
\item{x}{The vector to be modified}

\item{n}{The number of characters that the resulting vector should have}
}
\value{
A character vector of identical length to \code{x} in which all elements have n characters (or more, for those which already had more prior to processing)
}
\description{
Prepend one or more 0's to a number. Useful for alphabetizing facto levels named with numbers.
}
