% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_aggregate.R
\name{backup_aggregate}
\alias{backup_aggregate}
\title{Backup Aggregate}
\usage{
backup_aggregate(out_file = "backup_aggregate.zip")
}
\arguments{
\item{out_file}{The path of the file for writing the zip to}
}
\value{
A list
}
\description{
Create a backup of the entirety of the Aggregate server's data
}
