% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backup_central.R
\name{backup_central}
\alias{backup_central}
\title{Backup Central}
\usage{
backup_central(out_file = "backup.zip")
}
\arguments{
\item{out_file}{The path of the file for writing the zip to}
}
\value{
A list
}
\description{
Create a backup of the entirety of the Central server
}
