% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/credentials_check.R
\name{credentials_check}
\alias{credentials_check}
\title{Credentials check}
\usage{
credentials_check(credentials_path = NULL)
}
\arguments{
\item{credentials_path}{Path to a bohemia_credentials.yaml file}
}
\value{
A credentials file will be created and a bohemia_credentials environment variable will be set
}
\description{
Check on or create credentials. If a credentials_path argument is supplied, this function will set the bohemia_credentials environment variable to that path; if it is not supplied, this function will do the same, but will first prompt a few fields so as to create the credentials file.
}
