% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_form_list_aggregate.R
\name{get_form_list_aggregate}
\alias{get_form_list_aggregate}
\title{Get form list from aggregate}
\usage{
get_form_list_aggregate(pre_auth = FALSE)
}
\arguments{
\item{pre_auth}{Whether to pre-authenticate or not (some Aggregate versions require it)}
}
\value{
A list
}
\description{
Retrieve a dataframe of forms on the Aggregate server
}
