% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_form_schema.R
\name{get_form_schema}
\alias{get_form_schema}
\title{Get form schema}
\usage{
get_form_schema(fid)
}
\arguments{
\item{fid}{Form ID}
}
\value{
A dataframe
}
\description{
Retrieve a dataframe of a form schema from the Central server
}
