% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_any_object_from_aws.R
\name{retrieve_any_object_from_aws}
\alias{retrieve_any_object_from_aws}
\title{Retrieve any arbitrary object from the bohemia2022 bucket
Retrieve any arbitrary object from the bohemia2022 bucket}
\usage{
retrieve_any_object_from_aws(object, file)
}
\arguments{
\item{object}{Path to object to get}

\item{file}{Where to save locally}
}
\value{
A list
}
\description{
Retrieve any arbitrary object from the bohemia2022 bucket
Retrieve any arbitrary object from the bohemia2022 bucket
}
