% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_from_aws.R
\name{retrieve_data_from_aws}
\alias{retrieve_data_from_aws}
\title{Retrieve data from AWS}
\usage{
retrieve_data_from_aws(fid = NULL)
}
\arguments{
\item{fid}{Form ID for which data should be retrieved. If NULL, all}
}
\value{
A list
}
\description{
Retrieve data from the project's AWS S3 bucket
}
