% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_from_central.R
\name{retrieve_data_from_central}
\alias{retrieve_data_from_central}
\title{Retrieve data from central}
\usage{
retrieve_data_from_central(
  fids = NULL,
  except_fids = NULL,
  clean_column_names = TRUE,
  handle_sefull_split = TRUE
)
}
\arguments{
\item{fids}{Form IDs for which data should be retrieved. If NULL, all}

\item{except_fids}{Form IDs to exclude from retrieval}

\item{clean_column_names}{Whether to clean the column names}

\item{handle_sefull_split}{Whether to smartly handle the sefull split by combining all the sefull forms into one}
}
\value{
A list
}
\description{
Retrieve data from Central server
}
