% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_va_data_from_aws.R
\name{retrieve_va_data_from_aws}
\alias{retrieve_va_data_from_aws}
\title{Retrieve VA data from AWS}
\usage{
retrieve_va_data_from_aws()
}
\value{
A list
}
\description{
Retrieve VA data from the project's AWS S3 bucket. This is not for individual VA forms, but rather for the aggregated VA data as aggregated in the "pipeline" code
}
