% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_to_aws.R
\name{save_to_aws}
\alias{save_to_aws}
\title{Save to AWS}
\usage{
save_to_aws(robject = NULL, file_path = NULL, bucket_path = NULL, ...)
}
\arguments{
\item{robject}{An R object to save. Alternatively, use file_path}

\item{file_path}{The path to a file to save. Alternatively, use robject}

\item{bucket_path}{The name of the object to be saved in the bucket (including path and extension)}
}
\value{
Files will be written to AWS
}
\description{
Save data to AWS
}
