% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_values.R
\name{recode_values}
\alias{recode_values}
\title{Recode values}
\usage{
recode_values(
  data_list,
  schema_table,
  language = "english",
  variables_too = FALSE,
  messaging = TRUE,
  exclude_variables = c()
)
}
\arguments{
\item{data_list}{A list of data, like sefull$data}

\item{schema_table}{A dataframe of the form schema}

\item{language}{The language to be used. By default, "english"}

\item{variables_too}{Whether to recode variable names too. By default, FALSE}

\item{messaging}{Whether to show messages}

\item{exclude_variables}{A vector of variables to exclude from recoding (ie, those which are "select_multiple". Default: none)}
}
\value{
A dataframe
}
\description{
Given a list of data created by bohemiase::retrieve_data_from_aws and a "schema" table created by bohemiase::get_form_schema, recode values to their human-readable equivalents
}
