% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_from_aggregate.R
\name{retrieve_data_from_aggregate}
\alias{retrieve_data_from_aggregate}
\title{Retrieve data from aggregate}
\usage{
retrieve_data_from_aggregate(fids = NULL)
}
\arguments{
\item{fids}{Form IDs for which data should be retrieved. If NULL, all}
}
\value{
A list
}
\description{
Retrieve data from Aggregate server
}
