% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_from_aggregate.R
\name{retrieve_data_from_aggregate}
\alias{retrieve_data_from_aggregate}
\title{Retrieve data from aggregate}
\usage{
retrieve_data_from_aggregate(
  fids = NULL,
  except_fids = NULL,
  use_local_briefcase = TRUE,
  start_fresh = FALSE
)
}
\arguments{
\item{fids}{Form IDs for which data should be retrieved. If NULL, all}

\item{except_fids}{Form IDs to exclude from retrieval}

\item{use_local_briefcase}{Whether to use the local briefcase directory}

\item{start_fresh}{Whether to start data retrieval fresh (TRUE) or use already existant xml files in the briefcase directory (FALSE)}
}
\value{
A list
}
\description{
Retrieve data from Aggregate server
}
