% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_data_from_aws.R
\name{retrieve_data_from_aws}
\alias{retrieve_data_from_aws}
\title{Retrieve data from AWS}
\usage{
retrieve_data_from_aws(fid = NULL, central = TRUE, clean = FALSE)
}
\arguments{
\item{fid}{Form ID for which data should be retrieved. If NULL, all}

\item{central}{Whether to use the Central Server (by default, TRUE) or the aggregate server}

\item{clean}{Whether to retrieve clean data or not (by default, FALSE)}
}
\value{
A list
}
\description{
Retrieve ODK Central data from the project's AWS S3 bucket
}
