% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve_vatool_data_from_aws.R
\name{retrieve_vatool_data_from_aws}
\alias{retrieve_vatool_data_from_aws}
\title{Retrieve VAtool data from AWS}
\usage{
retrieve_vatool_data_from_aws()
}
\value{
A list
}
\description{
Retrieve VAtool data from the project's AWS S3 bucket
}
